#85 [fix] 행복 루틴 완료 로직 정리
